requirements.txt
alembic==1.10.2
click==8.1.3
colorama==0.4.6
dominate==2.7.0
Flask==2.2.3
Flask-Admin==1.6.1
Flask-Bootstrap4==4.0.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
visitor==0.1.3
Werkzeug==2.2.3
WTForms==3.0.1